\name{bootstrap_pvalues}
\alias{bootstrap_pvalues}
\title{ Bootstrap chi2 pvalues of association between a feature and a nominal attribute}
\description{ Bootstrap chi2 pvalues of association between a feature and a nominal attribute.}
\usage{ bootstrap_pvalues(x, B_max=1000,rseed=6013) }
\arguments{
   \item{x}{ The data matrix,}
   \item{B_max}{ The number of iteration of the bootstrap,}
   \item{rseed}{ The value to set the random number generator.}
}
\author{ Fabrice Colas }
