\name{data_PHARMAIT}
\alias{data_PHARMAIT}
\docType{data}
\title{ DATA set from Pharma-IT (Andreas Bender)}
\description{Not for public use.}
\usage{data(data_PHARMAIT)}
\format{ A data frame with 6181 observations on the following 59 variables.  }
\source{ Andreas Bender, Pharma-IT }
\keyword{datasets}
