\name{data_PHARMAIT_CLASSUNIQ}
\alias{data_PHARMAIT_CLASSUNIQ}
\docType{data}
\title{ DATA set from Pharma-IT with unique class column (Andreas Bender)}
\description{Not for public use. The same than \code{data_PHARMAIT_CLASSUNIQ}
but the profile of the molecules having an identical class name is averaged,
thus the 4472 observations instead of the 6181.}
\usage{data(data_PHARMAIT_CLASSUNIQ)}
\format{ A data frame with 4472 observations on the following 58 variables.  }
\source{ Andreas Bender, Pharma-IT }
\keyword{datasets}
