\name{init_data_cc}
\alias{init_data_cc}
\title{ Select observations without missing values on the variables involved of the cluster analysis }
\description{ The \code{set_canalysis_data} constructor does a systematic
initial preprocessing which involves selecting only those observations without
missing values on the variables for the cluster analysis. In the settings
(\code{attr(*,"settings")}), variables for the cluster analysis are defined by
the column (\code{in_canalysis}?). 

This initial transformation is necessary because model based clustering can not
process data sets with missing values. However, we do not drop observations
because of missing values on additional variables which are not involved in the
cluster analysis.  }
\usage{ init_data_cc(data, settings) }
\arguments{
  \item{data}{ a regular data matrix,}
  \item{settings}{ a configuration file as initially generated by
  \code{generate_canalysis_data_settings()} which, in a \code{canalysis_data}
  object is stored in \code{attr(*,"settings")}.}
}
\author{ Fabrice Colas }
\keyword{}
