\name{initPlot}
\alias{initPlot}
\title{ Plotting region initialization function}
\description{ This function initialize the device and the plotting region of the plot. }
\usage{ initPlot(cdata, figIdx, plotNbr=1, type="H") }
\arguments{ 
   \item{cdata}{ a \code{cdata} dataset, } 
   \item{figIdx}{ a figure index, } 
   \item{plotNbr}{ the number of graphics in a single plotting region, } 
   \item{type}{ the type of graphics.} 
 }
