\name{patternLowquant}
\alias{patternLowquant}
\title{Returns the 2.5\% quantile of the numeric vector}
\description{ Function that returns the 2.5\% quantile of a numeric vector,
where by default, the NA's are removed.  }
\usage{ patternLowquant(x) }
\arguments{ \item{x}{ A numeric vector. } }
\author{ Fabrice Colas }
\seealso{ \code{\link{help}}}
