\name{patternMax}
\alias{patternMax}
\title{Returns the maximum of the numeric vector}
\description{ Function that returns the maximum of a numeric vector,
where by default, the NA's are removed.  }
\usage{ patternMax(x) }
\arguments{ \item{x}{ A numeric vector. } }
\author{ Fabrice Colas }
\seealso{ \code{\link{help}}}
