\name{patternMean}
\alias{patternMean}
\title{Returns the mean of the numeric vector}
\description{ Function that returns the mean of a numeric vector,
where by default, the NA's are removed.  }
\usage{ patternMean(x) }
\arguments{ \item{x}{ A numeric vector. } }
\author{ Fabrice Colas }
\seealso{ \code{\link{help}}}
