\name{patternMedian}
\alias{patternMedian}
\title{Returns the median of the numeric vector}
\description{ Function that returns the median of a numeric vector,
where by default, the NA's are removed.  }
\usage{ patternMedian(x) }
\arguments{ \item{x}{ A numeric vector. } }
\author{ Fabrice Colas }
\seealso{ \code{\link{help}}}
\keyword{ }
