\name{patternSd}
\alias{patternSd}
\title{Returns the standard deviation of the numeric vector}
\description{ Function that returns the standard deviation of a numeric vector,
where by default, the NA's are removed.  }
\usage{ patternSd(x) }
\arguments{ \item{x}{ A numeric vector. } }
\author{ Fabrice Colas }
\seealso{ \code{\link{help}}}
