\name{patternUpquant}
\alias{patternUpquant}
\title{Returns the 97.5\% quantile of the numeric vector}
\description{ Function that returns the 97.5\% quantile of a numeric vector,
where by default, the NA's are removed.  }
\usage{ patternUpquant(x) }
\arguments{ \item{x}{ A numeric vector. } }
\author{ Fabrice Colas }
\seealso{ \code{\link{help}}}
