\name{plot.SDStability}
\title{Graphical report of an SDStability analysis}
\description{Illustrates by color images the result of stability analyzes (\code{SDStability}).}
\usage{plot.SDStability(x, ncolors=9, minmax=c(0.5, 1), ...)}
\arguments{\item{x}{ an \code{SDStability} object}
\item{ncolors}{ the number of colors in the grey gradient}
\item{minmax}{ the range of values for the color, by default \code{0.5} to \code{1} of the $V$ values.}
\item{\dots}{ (currently not use) additional parameters passed to the graphic functions}
}
\references{Colas, F, Meulenbelt, I, Houwing-Duistermaat, JJ, Kloppenburg, M, Watt, I, van Rooden, SM, Visser, M, Marinus, H, van Hilten, JJ, Slagboom, PE,  Kok, JN: Stability of Clusters for Different Time Adjustments in Complex Disease. Research 30th Annual International IEEE EMBS Conference (EMBC'08), Vancouver, Canada, 2008. [\url{http://dx.doi.org/10.1109/IEMBS.2008.4650238}]}
\author{Fabrice Colas}
\keyword{hplot}
\keyword{methods}
\alias{plot.SDStability}