\name{print.bicanalysis}
\alias{print.bicanalysis}
\title{ Generic printing function to show a bicanalysis}
\description{ For objects of type (\code{bicanalysis}), this generic
printing function returns a long table of the Bayesian Information Criterion
(BIC) scores of a subtype discovery analysis.}
\usage{ \method{print}{bicanalysis}(x, ...) }
\arguments{
   \item{x}{ An object of type \code{bicanalysis},}
   \item{...}{...}
}
\author{ Fabrice Colas }
\seealso{\code{\link{setBicanalysis}}}
