\name{print.canalysis_data}
\alias{print.canalysis_data}
\title{ Generic print function of a data set}
\description{ For objects of type (\code{canalysis_data}), this generic
print function just returns the main data set after being transformed.}
\usage{ print.canalysis_data(x,...) }
\arguments{
   \item{x}{ An object of type \code{canalysis_data},}
   \item{...}{ Additional generic arguments (NOTHING).}
}
\author{ Fabrice Colas }
\seealso{\code{\link{set_canalysis_data}}}
\keyword{}
