\name{print.cdata}
\alias{print.cdata}
\title{ Generic printing function for a dataset } 
\description{ For objects of type (\code{cdata}), this generic printing
function returns the data matrix result of the transformation.}
\usage{ \method{print}{cdata}(x, ...) }
\arguments{
   \item{x}{ An object of type \code{cdata},}
   \item{...}{...}
}
\author{ Fabrice Colas }
\seealso{\code{\link{setCdata}}}
