\name{print.cmodel}
\alias{print.cmodel}
\title{ Generic function to print a cluster model}
\description{ For objects of type (\code{cmodel}), this generic
printing function returns a cluster model. }
\usage{ \method{print}{cmodel}(x, ...) }
\arguments{
   \item{x}{ An object of type \code{cmodel},}
   \item{...}{...}
}
\author{ Fabrice Colas }
