\name{print.cresult}
\alias{print.cresult}
\title{ Generic print function with the characteristics of a cluster result}
\description{ For objects of type (\code{cresult}), this generic
print function reports cross comparison of the most likely cluster results
(\code{cross_compare_models}).}
\usage{ print.cresult(x, html =TRUE, ...) }
\arguments{
   \item{x}{ An object of type \code{cresult},}
   \item{html}{ A logicial value (\code{...}) telling whether we should generate an HTML output. }
   \item{...}{...}
}
\author{ Fabrice Colas }
\seealso{\code{\link{cross_compare_models}}, \code{\link{set_cresult}}}
\keyword{}
