\name{SDStability}
\title{SDStability}
\description{SDStability}
\usage{SDStability(x, q, rseed=6013, nnoise=10, nrep=10)}
\arguments{\item{x}{ an \code{SDisc} analysis result}
\item{q}{ the model to assess}
\item{rseed}{ an integer to initialize the random number generator}
\item{nnoise}{ an integer to set the number of noise levels }
\item{nrep}{ an integer to indicate the number of replicates}
}
\references{Colas, F, Meulenbelt, I, Houwing-Duistermaat, JJ, Kloppenburg, M, Watt, I, van Rooden, SM, Visser, M, Marinus, H, van Hilten, JJ, Slagboom, PE,  Kok, JN: Stability of Clusters for Different Time Adjustments in Complex Disease. Research 30th Annual International IEEE EMBS Conference (EMBC'08), Vancouver, Canada, 2008. [\url{http://dx.doi.org/10.1109/IEMBS.2008.4650238}]}
\author{Fabrice Colas}
\keyword{robust}
\alias{SDStability}