\name{setup_PHARMAIT_data}
\alias{setup_PHARMAIT_data}
\docType{data}
\title{Data setup file for the PharmaIT analysis}
\description{ Given a data set, a basic settings file can be generated from
\code{generate_canalysis_data_settings} and it should be edited then.}
\seealso{\code{\link{generate_canalysis_data_settings}}}
\usage{data(setup_PHARMAIT_data)}
\keyword{datasets}
