\name{SubtypeDiscovery-package}
\alias{SubtypeDiscovery-package}
\alias{SubtypeDiscovery}
\docType{package}
\title{
What the package does (short line)
~~ titre du package ~~
}
\description{
More about what it does (maybe more than one line)
~~ Description concise (1-5 lignes) du package ~~
}
\details{
\tabular{ll}{
Package: \tab SubtypeDiscovery\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2008-05-06\cr
License: \tab What license is it under?\cr
}
~~ Une vue d'ensemble montrant comment utiliser le package, y compris ses ~~
~~ fonctions les plus importantes ~~
}
\author{
Who wrote it

Responsable : Who to complain to <yourfault@somewhere.net>
~~ L'auteur et/ou la personne responsable du package ~~
}
\references{
~~ Références bibliographiques ou autre références pour de l'information générale ~~
}
~~ Eventuellement d'autres mots clés, un par ligne, issus du fichier ~~
~~ KEYWORDS du répertoire documentation de R ~~
\keyword{ package }
\seealso{
~~ Liens valides vers d'autres pages de manuel, par exemple ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
}
