\name{summary.cdata}
\alias{summary.cdata}
\title{ Generic function summarizing a data set}
\description{ For objects of type (\code{cdata}), this generic
summarizing function describes a number of general information and
characteristics on the dataset.}
\usage{ \method{summary}{cdata}(object, shift="", ...) }
\arguments{
   \item{object}{ an object of type \code{cdata},}
   \item{shift}{ a left-shift in spaces when called from \code{summary.cresult},}
   \item{...}{...}
}
\author{ Fabrice Colas }
