\name{summary.cresult}
\alias{summary.cresult}
\title{ Generic function characterizing graphically a data set}
\description{ For objects of type (\code{cresult}), this generic summary
function describes the main information, characteristics and results of a
subtype discovery analysis.}
\usage{ \method{summary}{cresult}(object, type='term', ...) }
\arguments{
   \item{object}{ An object of type \code{cresult},}
   \item{type}{ in \code{html}, \code{latex}, or \code{term},}
   \item{...}{...}
}
\author{ Fabrice Colas }
