\name{to_binary}
\alias{to_binary}
\title{ Convert a matrix to a binary representation with respect to the median of each feature}
\description{Convert a matrix to a binary representation with respect to the median of each feature.}
\usage{ to_binary(x) }
\arguments{
   \item{x}{ The data matrix.}
}
\author{ Fabrice Colas }
\seealso{\code{\link{association_analysis}}}
