\name{toBinary}
\alias{toBinary}
\title{ Convert a matrix to a binary representation with respect to the median of each feature}
\description{Convert a matrix to a binary representation with respect to the median of each feature.}
\usage{ toBinary(x) }
\arguments{
   \item{x}{ The data matrix.}
}
\author{ Fabrice Colas }
\seealso{\code{\link{associationAnalysis}}}
