\name{varFeatAnalysis}
\alias{varFeatAnalysis}
\title{ in development}
\description{ in development }
\usage{ varFeatAnalysis(m, a=NULL, maxFeat=NA, postscript=TRUE,clusterModel='VVI', clusterG=6,
   clusterRseed=6013:6063) }
\arguments{ 
   \item{m}{ data matrix,} 
   \item{a}{ association analysis result,} 
   \item{maxFeat}{ maximum number of features,} 
   \item{postscript}{ whether a postscript file should be produced,} 
   \item{clusterModel}{ the type of covariance model,} 
   \item{clusterG}{ the number of subtypes,} 
   \item{clusterRseed}{ the different random initialization numbers.} 
 }
\author{ Fabrice Colas }
