\name{wada2005}
\alias{wada2005}
\docType{data}
\author{Edward O. Cannon}
\title{ Molecule descriptors of prohibited substance from the World Anti Doping Agency (WADA, 2005)}
\description{ Originally generated by Edward O. Cannon [1,2,3].

All 249,071 three dimensional CORINA[4] generated structures were taken from
the publicly available 1999 National Cancer Institute database [5]. Duplicates
were removed, leaving 236,936 unique structures, which were then filtered for
drug-likeness, using a Lipinski filter [6] in MOE [7], leaving 111,694
structures. A further 463 metal ion complexes were filtered from the database
as no bits were set in the MACCS key descriptor for these compounds. This left
a dataset of 111,231 NCI compounds. These compounds were combined with the
original 5,245 compounds taken from the 2005 WADA Anti-Doping Agency (WADA)
dataset [8] to form a final dataset of 116,476 compounds. The WADA classes are
composed of molecules explicitly on the prohibited list and of molecules of
similar biological activity and chemical structure taken from the MDDR database
[9] (version 2003.1). The WADA dataset is composed of 10 different activity
classes:

\item beta blockers (P2), 
\item anabolic agents (S1), 
\item hormones and related substances (S2), 
\item $\beta$-2 agonists (S3), 
\item agents with anti-estrogenic activity (S4), 
\item diuretics and other masking agents (S5), 
\item stimulants (S6), 
\item narcotics (S7),
\item cannabinoids (S8) and 
\item glucocorticosteroids (S9). 

The breakdown of the WADA activity classes is given in Table 1 of [1] and in
the supplementary information (Additional file 1 in [1]). Pictures of the most
and least representative molecules for each prohibited class can also be found
in the supplementary information (see Additional files 2 and 3 in [1]). For the
purpose of this work, the NCI compounds were assumed to be inactive, with no
NCI compound being present in the WADA dataset after an initial screen that
compared canonical SMILES strings.  }
\usage{data(wada2005)}
\format{ A data frame of 5,245 compounds }
\source{ 
[1] Cannon, E.O., Nigsch, F., Mitchell, J.B.O.: A novel hybrid ultrafast shape
descriptor method for use in virtual screening. Chemistry Central Journal 2
(2008) 9 

[2] Cannon, E.O., Mitchell, J.B.O.: Classifying the world anti-doping agency?s
2005 prohibited list using the chemistry development kit ?ngerprint. [3]
173?182 

[3] Cannon, E.O., Bender, A., Palmer, D.S., Mitchell, J.B.O.M.B.O.:
Chemoinformatics-based classi?cation of prohibited substances employed for
doping in sport. Journal of Chemical Information and Modeling 46(6) (2006)
2369?2380

[4] CORINA [http://www.mol-net.de]

[5] The National Cancer Institute Database
[http://cactus.nci.nih.gov/ncidb2/download.html]

[6]  Lipinski CA, Lombardo F, Dominy BW, Feeny PJ:  Experimental and
Computational Approaches to Estimate Solubility and Permeability in Drug
Discovery and Development Settings.  Adv Drug Deliv Rev 1997, 23:3-25.

[7] Chemical Computing Group: Suite 910, 1010 Sherbrooke St. W, Montreal,
Quebec H3A 2R7, Canada. 

[8] World Anti-Doping Agency (WADA): Stock Exchange Tower, 800 Place Victoria
(Suite 1700), PO Box 120, Montreal (Quebec) H4Z 1B7, Canada. 

[9]  Elsevier MDL: 2440 Camino Ramon, Suite 300, San Ramon, CA 94583. 

}
\keyword{datasets}
