\name{wada2008}
\alias{wada2008}
\docType{data}
\author{Edward O. Cannon}
\title{ Molecular descriptors of prohibited substances taken from the World
Anti-Doping Agency's 2008 Prohibited List}
\description{ Originally generated by Edward O. Cannon in 2008. The dataset is
similar to the wada2005 dataset which was previously published in [1,2,3]. See
below for the description of the 2008 dataset.

The 2008 WADA (World Anti-Doping Agency) [4] dataset is broken down into 10
different activity classes:

\item beta blockers (P2), 
\item anabolic agents (S1), 
\item hormones and related substances (S2), 
\item $\beta$-2 agonists (S3), 
\item hormone antagonists and modulators (S4), 
\item diuretics and other masking agents (S5), 
\item stimulants (S6), 
\item narcotics (S7),
\item cannabinoids (S8) and 
\item glucocorticosteroids (S9). 

The dataset is composed of substances taken from the 2008 Prohibited List [5]
and molecules with similar biological activity and chemical structure taken
from the MDDR (MDL Drug Data Report) database (Version 2003.1) [6]. The MDDR
ids are listed below:

\item 12352     Cannabinoid Agonist
\item 16100     Adrenergic (beta) Blocker,
\item 31250     Adrenergic (beta) Blocker,
\item 31251     Adrenergic (beta 1) Blocker,
\item 34000     Diuretic,
\item 37320     Plasminogen Activator,
\item 39100     Glucocorticoid,
\item 41300     Estrogen Receptor Modulator,
\item 42100     Growth Hormone,
\item 42310     Corticotropin Releasing Factor,
\item 42500     Gonadotropin,
\item 43110     Insulin,
\item 43136     Adrenoceptor (beta 3) Agonist,
\item 43139     Insulin-Like Growth Factor,
\item 75711     Anti-Estrogen,
\item 75721     Aromatase Inhibitor,
\item 78451     Steroid (5-alpha) Reductase 1-Inhibitor.

All molecules were imported into MOE (Molecular Operating Environment, Version
2005.06) [7], washed to remove counterions and energy minimized using the
MMFF94 force field [8]. All 184 two dimensional descriptors were calculated. 

}
\usage{data(wada2008)}
\format{ A data frame of 3,038 compounds }
\seealso{\code{\link{wada2008_settings}}}
\source{ 
[1] Cannon, E.O., Nigsch, F., Mitchell, J.B.O.: A novel hybrid ultrafast shape
descriptor method for use in virtual screening. Chemistry Central Journal 2
(2008) 3. 

[2] Cannon, E.O., Mitchell, J.B.O.: Classifying the world anti-doping agency's
2005 prohibited list using the chemistry development kit fingerprint. Lecture
Notes in Bioinformatics 4216 (2006) 173-182.  

[3] Cannon, E.O., Bender, A., Palmer, D.S., Mitchell, J.B.O.:
Chemoinformatics-based classification of prohibited substances employed for
doping in sport. Journal of Chemical Information and Modeling 46(6) (2006)
2369-2380.

[4] World Anti-Doping Agency (WADA): Stock Exchange Tower, 800 Place Victoria
(Suite 1700), PO Box 120, Montreal (Quebec) H4Z 1B7, Canada. 

[5] Prohibited List
[http://www.wada-ama.org/rtecontent/document/2008\_List\_En.pdf]
Accessed: 03/06/08.

[6]  Elsevier MDL: 2440 Camino Ramon, Suite 300, San Ramon, CA 94583. 

[7] Chemical Computing Group: 1010 Sherbrooke St. W, Suite 910
Montreal, Quebec H3A 2R7, Canada.

[8] Halgren, T. A.: Merck Molecular Force Field. 1. Basis, form, scope,
parameterization, and performance of MMFF94. Journal of Computational Chemistry
17 (1996) 490-519.  }
\keyword{datasets}
