\name{wada2008_settings}
\alias{wada2008_settings}
\docType{data}
\author{Fabrice Colas, Ed O. Cannon}
\title{Settings for the WADA 2008 data set}
\description{Settings for the WADA 2008 data set. Together with the data set
(\code{wada2008}), to be provided as argument to (\code{set_cdata}). }
\usage{data(wada2008_settings)}
\seealso{\code{\link{wada2008}}}
\examples{ data(wada2008_settings) }
\keyword{datasets}
