\name{analysis}
\alias{analysis}
\title{Sample workflow-method for subtype discovery}
\description{ Workflow example for subtype discovery by cluster analysis. This
workflow generates the graphics that illustrate the selected data
transformation.  Then, it proceeds to a cluster analysis. Further, we
characterize the cluster results by table-reports. And we finally cross-compare
the top most likely results.  }
\usage{ analysis(x) }
\arguments{ 
\item{x}{Of type \code{cresult} which can be constructed from the
   method \code{set_cresult}, } 
 }
\author{ Fabrice Colas }
\keyword{ cluster }
\keyword{ models }
\keyword{ design }
\keyword{ manip }
\keyword{ hplot }
