\name{get_best_models}
\alias{get_best_models}
\title{ Get the list of top ranking models in a cresult }
\description{ Returns a character vector of the type "VVV,3" that lists the top
ranking models in the \code{cresult}. }
\usage{ get_best_models(x) }
\arguments{
  \item{x}{ A cluster analysis data structure of type (\code{cresult})}. }
\value{ A character vector with elements of the type "modelName, number of
cluster".  }
\author{ Fabrice Colas }
\keyword{}
