\name{SubtypeDiscovery-package}
\alias{SubtypeDiscovery-package}
\alias{SubtypeDiscovery}
\docType{package}
\title{
Selection of methods to characterize, compare and evaluate cluster models for medical research.
}
\description{
In order to identify homogeneous clusters we construct visual-aids and
table-charts that summary the key-characteristics of the different clusters for
the different settings of the clustering algorithm. We also provide a strategy
to further cross-compare the top cluster results. Finally, to evalute cluster
results in medical research on sibling pairs, we especially report two
statistical tests measuring familial aggregation. 
}
\details{
\tabular{ll}{
Package: \tab SubtypeDiscovery\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2008-04-28\cr
License: \tab BSD\cr
}


}
\author{
Fabrice Colas, Ingrid Meulenbelt, Jeanine Houwing-Duistermaat, Stephanie van Rooden, Martine Visser

Responsable : Fabrice Colas <fcolas@liacs.nl>
}
\references{
~~ Références bibliographiques ou autre références pour de l'information générale ~~
}
~~ Eventuellement d'autres mots clés, un par ligne, issus du fichier ~~
~~ KEYWORDS du répertoire documentation de R ~~
\keyword{ clustering, subtype discovery, cross-comparison, visual-aid, table-chart, medical research, familial aggregation }
%%\seealso{
%~~ Liens valides vers d'autres pages de manuel, par exemple ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%}
\examples{
}
