\name{SubtypeDiscovery-package}
\alias{SubtypeDiscovery-package}
\alias{SubtypeDiscovery}
\docType{package}
\title{Methods for Characterization, Comparison and Evaluation of Cluster
Models}
\description{ We aim to identify subtypes of diseases that present clinical
heterogeneity. We do so by  searching for homogeneous clusters  in values of
markers that reflect the severity of the disease.  As we cluster data by models
that use different  spatial-shapes, we  need a method that, first, reports the
essential characteristics of each cluster result and second, that offers a
comparative view between cluster results. On top of that, as we analyse data
from medical research, we also consider the clinical relevance of the clusters
as an important feature. Therefore, in the current package, we consider three
important items. First, how do we \emph{characterize} efficiently different
clusters of a cluster result? Second, how do we \emph{compare}  different
cluster results? And third, how do we \emph{evaluate} clusters for their
clinical relevance?  We answer these  questions, by creating  visual-aids and
table-charts that show the  characteristics of  clusters and models under these
different views.  }
\details{
\tabular{ll}{
Package: \tab SubtypeDiscovery\cr
Type: \tab Package\cr
Version: \tab 1.8\cr
Date: \tab 2008-05-23\cr
License: \tab To be defined. All rigths reserved.\cr
}

}
\author{ Fabrice Colas

Responsable : Fabrice Colas <fcolas@liacs.nl> }
\references{
Hastie, T., Tibshirani, R., Friedman, J.: The Elements of Statistical Learning,
Data Mining, Inference, and Prediction. Springer Series in Statistics. Springer
(2001) 

Sneath, P.H.A., Sokal, R.R.: Numerical Taxonomy, The Principles and Practice of
Numerical Classification. Books in Biology. W. H. Freeman and Company (1973) 

Beyer, K., Goldstein, J., Ramakrishnan, R., Shaft, U.: When is 'nearest
neighbor' meaningful? Lecture Notes in Computer Science 1540 (1999) 217?235 

Fraley, C., Raftery, A.E.: MCLUST: Software for model-based cluster analysis.
Journal of Classi?cation 16 (1999) 297-306 

Fraley, C., Raftery, A.E.: Model-based clustering, discriminant analysis and
density estima- tion. Journal of the American Statistical Association 97 (2002)
611-631 

Fraley, C., Raftery, A.E.: Enhanced software for model-based clustering,
density estimation, and discriminant analysis: MCLUST. Journal of Classi?cation
20 (2003) 263-286 

Fraley, C., Raftery, A.E.: MCLUST version 3 for R: Normal mixture modeling and
model- based clustering. Technical Report 504, University of Washington,
Department of Statistics (September 2006) 

Banfield, J.D., Raftery, A.E.: Model-based Gaussian and non-Gaussian
clustering. Biometrics 49 (1993) 803-821 

Kass, R.E., Raftery, A.E.: Bayes factors. Journal of the American Statistical
Association 90(430) (1995) 773-? 

Eisen, M.B., Spellman, P.T., Brown, P.O., Botstein, D.: Cluster analysis and
display of genome-wide expression patterns. In: Proceedings of National Academy
of Science USA.  Volume 95. (1998) 11863-14868 

Inselberg, A.: The plane with parallel coordinates. The Visual Computer 1(2)
(1985) 69?91 Tukey, J.W.: Exploratory Data Analysis. MA: Addison-Wesley. (1977) 

Tufte, E.R.: The Visual Display of Quantitative Information. Graphics Press,
Cheshire, Connectictut Connectictut (1983) 

Tufte, E.R.: Envisioning Information. Graphics Press, Cheshire, Connectictut
(1990) 

Brewer, C.A.: 7. In: Color Use Guidelines for Mapping and Visualization.
Elsevier Science, Tarrytown, NY (1994) 123-147 

Ewens, W.J., Grant, G.R.: Statistical Methods in Bioinformatics, An
Introduction. Springer (2005) 

Gentleman, R., Carey, V.J., Huber, W., Irizarry, R.A., Dudoit, S.:
Bioinformatics and Computational Biology Solutions Using R and Bioconductor.
Springer Series in Statistics for Biology and Health. Springer (2005)
}

\keyword{ aplot }
\keyword{ dplot }
\keyword{ hplot }
\keyword{ manip }
\keyword{ classes }
\keyword{ models }
\keyword{ multivariate }
\keyword{ cluster }
\keyword{ package }

\seealso{
\code{\link[mclust:mclust-package]{mclust}} 
\code{\link[RColorBrewer:RColorBrewer-package]{RColorBrewer}} 
\code{\link[e1071:e1071-package]{e1071}} 
}
