\name{SubtypeDiscovery-package}
\alias{SubtypeDiscovery-package}
\alias{SubtypeDiscovery}
\docType{package}
\title{
Selection of methods to characterize, compare and evaluate cluster models for medical research.
}
\description{
In order to identify homogeneous clusters we construct visual-aids and
table-charts that summary the key-characteristics of the different clusters for
the different settings of the clustering algorithm. We also provide a strategy
to further cross-compare the top cluster results. Finally, to evalute cluster
results in medical research on sibling pairs, we especially report two
statistical tests measuring familial aggregation. 
}
\details{
\tabular{ll}{
Package: \tab SubtypeDiscovery\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2008-04-29\cr
License: \tab BSD\cr
}
The entry-function is analysis which allows multiple parameters that specify the type of analysis to carry. 
}
\author{
Fabrice Colas <fcolas at liacs.nl
}
\references{
\emph{Selected Methods for Characterization, Comparison and Evaluation of Cluster Models for Medical Research}, unpublished F. Colas et al.

\emph{Stability of Clusters for Different Time Adjustments in Complex Disease Research}, under review for IEEE EMBC 2008, F. Colas et al.
}
\keyword{ package }
\seealso{
\code{\link[<pkg>:<pkg>-package]{<pkg>}} 
}
\examples{
exemples simples des fonctions les plus importantes 
}
