\name{texTable}
\title{texTable}
\description{texTable}
\usage{texTable(x, cap, lab="tab:", digits=NULL, sanitize=TRUE, ...)}
\arguments{\item{x}{ a matrix}
\item{cap}{ the caption of the LaTeX table}
\item{lab}{ the label of the LaTeX table}
\item{digits}{ the number of digits to display}
\item{sanitize}{ whether to sanitize the row and column names values for LaTeX table (\code{xtable}). If so, define an alternative treatment function, }
\item{\dots}{ additional parameters to be passed to \code{xtable}}
}
\author{Fabrice Colas}
\seealso{\code{\link{print.bicTable}}, \code{\link{print.SDData}}, \code{\link{print.SDisc}}, \code{\link{SDDataSettings}}, \code{\link{summary.bicTable}}, \code{\link{summary.SDCModel}}, \code{\link{summary.SDData}}, \code{\link{summary.SDisc}}, }\keyword{manip,print}
\alias{texTable}