\name{today}
\alias{today}
\title{Today date as 'YYYY-MM-DD'}
\description{ Return today date string as 'YYYY-MM-DD', e.g. 2008-02-31. Useful
to track files generated in the course of the different analysis.}
\usage{ today() }
\author{ Fabrice Colas }
\seealso{ \code{\link{paste2}}}
\examples{ print(today()) }
\keyword{}
